+++ src/segcraft/__init__.py
@@ -1,6 +1,14 @@
 """SegCraft package."""
 
-from .api import evaluate, load_config, load_config_object, predict, train
+from .api import evaluate, list_available_presets, load_config, load_config_object, predict, train
 from .config import SegCraftConfig
 
-__all__ = ["SegCraftConfig", "load_config", "load_config_object", "train", "evaluate", "predict"]
+__all__ = [
+    "SegCraftConfig",
+    "evaluate",
+    "list_available_presets",
+    "load_config",
+    "load_config_object",
+    "predict",
+    "train",
+]